+++ Front-End/Styles/qr.css
@@ -42,6 +42,8 @@ body,
 /* Create an active/current tablink class */
 .tab button.active {
     background-color: #ccc;
+    background-color: #00A1E0;
+    color: white;
 }
 
 /* Style the tab content */
